+++ dbm-ui/frontend/src/views/db-manage/mysql/MYSQL_CLIENT_CLONE_RULES/components/SourceColumn.vue
@@ -47,6 +47,7 @@
     :label="t('模块')"
     :loading="loading"
     :min-width="150"
+    readonly
     required>
     <EditableBlock
       v-model="module"
