+++ dbm-ui/frontend/src/views/db-manage/mysql/MYSQL_PROXY_ADD/Index.vue
@@ -156,9 +156,9 @@
         id: 0,
         master_domain: '',
         region: '',
-        related_clusters: [] as RowData['cluster']['related_clusters'],
-        spec_id_list: [] as RowData['cluster']['spec_id_list'],
-      },
+        related_clusters: [],
+        spec_id_list: [],
+      } as RowData['cluster'],
       data.cluster,
     ),
     labels: (data.labels || []) as RowData['labels'],
