+++ dbm-ui/frontend/src/views/db-manage/mysql/MYSQL_PROXY_SWITCH/Index.vue
@@ -59,14 +59,16 @@
         ref="table"
         :source-type="sourceType"
         :ticket-details="ticketDetails" />
-      <BkFormItem
-        v-bk-tooltips="t('存在业务连接时需要人工确认')"
-        class="fit-content">
+      <BkFormItem>
         <BkCheckbox
           v-model="formData.force"
           :false-label="false"
           true-label>
-          <span class="safe-action-text">{{ t('检查业务连接') }}</span>
+          <span
+            v-bk-tooltips="t('存在业务连接时需要人工确认')"
+            class="safe-action-text">
+            {{ t('检查业务连接') }}
+          </span>
         </BkCheckbox>
       </BkFormItem>
       <TicketPayload v-model="formData.payload" />
