+++ dbm-ui/frontend/src/views/db-manage/tendb-cluster/TENDBCLUSTER_INSTANCE_CLONE_RULES/components/SourceColumn.vue
@@ -43,6 +43,7 @@
     :label="t('所属集群')"
     :loading="loading"
     :min-width="250"
+    readonly
     required>
     <EditableBlock
       v-model="clusterDomain"
