+++ dbm-ui/frontend/src/views/db-manage/tendb-cluster/TENDBCLUSTER_NODE_REBALANCE/Index.vue
@@ -107,14 +107,14 @@
   import TicketPayload, {
     createTickePayload,
   } from '@views/db-manage/common/toolbox-field/form-item/ticket-payload/Index.vue';
+  import ClusterColumn from '@views/db-manage/tendb-cluster/common/toolbox-field/cluster-column/Index.vue';
 
   import { random } from '@utils';
 
   import CapacityColumn from './components/capacity-column/Index.vue';
-  import ClusterColumn from './components/ClusterColumn.vue';
 
   interface RowData {
-    cluster: ComponentProps<typeof ClusterColumn>['modelValue'];
+    cluster: TendbClusterModel;
     labels: ComponentProps<typeof ResourceTagColumn>['modelValue'];
     targetCapacity: ComponentProps<typeof CapacityColumn>['modelValue'];
   }
@@ -150,7 +150,7 @@
         master_domain: '',
         region: '',
         remote_shard_num: 0,
-      },
+      } as unknown as TendbClusterModel,
       data.cluster,
     ),
     labels: (data.labels || []) as RowData['labels'],
